refactor: Extract zoom and font size logic into a separate module
Separated the zoom and font size functionality from the main editor initialization file, placing it into its own module "zoomLogic". This helps to make the code base more organized and maintainable. The refactoring encourages separation of concerns and improves code readability
